Made the LRU cache thread-safe (hopefully).
How it will perform is totally up in the air at this point.  But this
paves the way for some other experimentation I want to do.

+++ config.hpp
@@ -9,7 +9,7 @@ extern float32 dice_rate;
 extern float32 focus_factor;
 extern float32 min_upoly_size;
 extern int max_grid_size;
-extern int grid_cache_size;
+extern float grid_cache_size;
 
 // Statistics
 extern uint64 split_count;
